+++ src/parser/btor2/parser.h
@@ -41,6 +41,8 @@ class Parser : public bzla::parser::Parser
 
   bool parse_term(const std::string& input, bitwuzla::Term& res) override;
   bool parse_sort(const std::string& input, bitwuzla::Sort& res) override;
+  std::vector<bitwuzla::Sort> get_declared_sorts() const override;
+  std::vector<bitwuzla::Term> get_declared_funs() const override;
 
  private:
   enum class ParsedKind